feat: Add custom serializer for BlogPage model
This commit introduces a custom serializer for the BlogPage model in a new file called serializers.py, inside the blog/api folder. This custom serializer called BlogPageSerializer overwrites the "to_representation" method in order to include the 'intro' field when representating BlogPage instances through an API view. This way, the 'intro' field is always included when querying the database for a blog post and removes the need of specifying this field in the URL as a parameter, like "api/v2/posts/?fields=intro".